requirements.txt
PyYAML
torch>=1.9.0
numpy>=1.17
scipy>=1.2.1
prefetch_generator>=1.0.1
msgpack>=0.5.6
sortedcontainers
uvloop>=0.14.0
grpcio-tools>=1.33.2
protobuf>=5.29.0
configargparse>=1.2.3
cryptography>=3.4.6
pydantic>=2.0.0
packaging>=20.9
varint>=1.0.2
netaddr>=1.3.0
idna>=3.10
py-multibase>=1.0.3
transformers>=4.43.1
morphys>=1.0
py-multicodec>=0.2.1
tenacity>=8.2.3
python-dotenv>=1.0.1
substrate-interface==1.7.11
humanfriendly
fastapi>=0.115.13
aiohttp>=3.12.13
uvicorn>=0.34.3
prometheus-client>=0.22.1
